+++ docs/changelog.md
@@ -4,6 +4,7 @@ icon: material/alert-decagram
 
 #### 1.12.13
 
+* Fix naive inbound
 * Fixes and improvements
 
 #### 1.12.12